requirements.txt
ricecooker>=0.6.20
PyYAML==3.12
